requirements.txt
Django==1.11.8
djangorestframework==3.7.3
